continue to extract fold.rs
